feat(tagsInput): Add shouldAddTag/shouldRemoveTag.
Add an option for the user to set if a tag should be created.
This feature will allow users to define their own custom rules
for when a tag can be added.

Add an option for the user to set if a tag should be removed.
This feature will allow users to define their own custom rules
for when a tag can be removed.

Closes #100